requirements.txt
attr==0.3.2
ConfigParser==6.0.0
cryptography==42.0.3
Cython==3.0.8
Django==4.2.1
Django==4.2.9
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.13.2
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
numpy==1.26.4
pandas==2.2.0
Pillow==9.5.0
Pillow==10.2.0
protobuf==4.25.3
pyOpenSSL==24.0.0
pywin32==306
railroad==0.5.0
scikit_learn==1.2.2
scikit_learn==1.3.2
Sphinx==7.2.6
thread==0.1.3
tornado==6.4
trove_classifiers==2024.1.31
truststore==0.8.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
